Improve performance of REPEAT functions
